Re-organization of exceptions
