+++ src/testplates/impl/validators/enum.py
@@ -14,7 +14,7 @@ from resultful import (
     Result,
 )
 
-from testplates.impl.base import (
+from testplates.impl.exceptions import (
     TestplatesError,
 )
