+++ src/testplates/impl/validators/utils.py
@@ -15,7 +15,7 @@ from resultful import (
     Result,
 )
 
-from testplates.impl.base import (
+from testplates.impl.exceptions import (
     TestplatesError,
 )
